requirements.txt
google-generativeai
pyTelegramBotAPI
python-dotenv
spotipy
